Several improvements to viewmesh: - Added background loading of the object. - Removed the box world, it's more of an obstruction than helpful. - Fixed lighting. - Changed default camera mode to rotate. - Locked camera movement until an object is loaded. - Reset the camera when a new object is loaded. - GUI scaling fixes. - Fixed a few crashes. - A few warnings fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33797 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/viewmesh/genmeshasset.cpp
@@ -121,7 +121,7 @@ csPtr<iStringArray> GenmeshAsset::GetSockets()
 
   for (size_t i = 0; i < skeleton->GetFactory()->GetSocketsCount(); i++)
   {
-    iSkeletonSocketFactory* sock = skeleton->GetFactory()->GetSocket(i);
+    iSkeletonSocketFactory* sock = skeleton->GetFactory()->GetSocket((int)i);
     if (!sock) continue;
 
     csString s = sock->GetName();
@@ -228,7 +228,7 @@ bool GenmeshAsset::DetachAll()
   if (!skeleton) return false;
   for (size_t i = 0; i < skeleton->GetFactory()->GetSocketsCount(); i++)
   {
-    iSkeletonSocketFactory* sock = skeleton->GetFactory()->GetSocket(i);
+    iSkeletonSocketFactory* sock = skeleton->GetFactory()->GetSocket((int)i);
     if (!sock) continue;
 
     AttachMesh(sock->GetName(), 0);
